웨폰데이터 변경
무기를 강화한 데이터가 게임이 종료 된 후에도 남아있어서 이를 수정했습니다

+++ gamemakejo/Assets/Scripts/BossMonster.cs
@@ -8,7 +8,7 @@ public class BossMonster : MonoBehaviour
     private GameObject _target = null;
 
     [SerializeField] private int _damage = 3;
-    public float HP = 10;
+    public int HP = 10;
     private void OnCollisionEnter(Collision coll)
     {
         if (coll.gameObject.CompareTag("Player"))
